Normative: support Etc/Unknown time zone
This PR adds explicit support for an Etc/Unknown time zone ID, which
ICU returns when the system time zone is not recognized, usually because
the host environment's OS uses an ID from a newer version of the IANA
Time Zone Database that the ECMAScript engine has not yet been
upgraded to.

This time zone should behave identically to UTC, except it should have
a different localized name to help programmers and end users realize
that there's a problem that requires someone to resolve.

See https://github.com/tc39/proposal-canonical-tz/pull/25 for a previous
attempt to address this issue.